Read from pipes on Windows and return system errors
 - Using ReadFile rather than ReadConsoleW allows you to read from
pipes.  In addition, the documentation for gopass specifies that
UTF-8 is not supported so there is no reason to specifically try
and read Unicode from the console.

 - Properly inspects the return value from the Windows syscall to
see if there was an error or not and returns it. (Non-zero return
value indicates success; the error itself is always non-nil and
would reflect some information about the successful state if there
was not an error).