Handle errors first and disregard byte when error exists
Fixes a problem where error cases were usually simply treated
as `enter` presses on Windows and nil bytes on *nix.  This was
a result of the error case being handled last rather than first.

Any byte that comes back along with an error should be disregarded.

Added test cases that confirm this behavior.