Remove os variation; use ssh/terminal for necessary terminal calls
 - The golang/crypto/ssh/terminal package properly makes terminals
raw on bsd, linux, and windows.  There is no need to make the
syscalls ourselves now for terminal handling.
 - The ReadFile syscall on Windows is just the underlying
implementation for *File.Read(); using the higher level calls
allows all OSs to share an implementation. The same goes for
the *nix system calls.  File reading can be done centrally as
well.
 - Confirmed tests pass on OSX and Windows manually.
 - Moved terminal handling into the getPasswd method now that
it is not OS dependend to remove unnecessary calls.
 - Fixes an issue where gopass would fail to pipe input properly due
to inability to make raw.  Now tests check that os.Stdin is a terminal
before trying to make raw.
 - Adds tests to ensure proper pipe behavior
 - Modifies travis to test on linux and osx