Add support for Illumos-derived operating systems
This commit adds support for building on Illumos-derived operating
systems by reimplementing the various missing functions from package
golang.org/x/crypto/ssh/terminal when compiling for the "solaris" GOOS.

The implementations are based on the C in the Illumos libraries. In
order to support calling either version of the functions, they are
abstracted into a separate file with build tags, where the "non-solaris"
version delegates to the crypto/ssh/terminal functions.

Regrettably the crypto/ssh/terminal functions return a type with private
fields which our private implementation cannot construct  - consequently
we return an implementation dependent struct from makeRaw and use that
for the call to restore.